[clang-tidy] bugprone-header-guard : a general version of llvm-header-guard
A general use, missing header guard finder and fixer.
No style checks.  If there is already a working header guard, it is used.
If the header guard is missing or broken, it is created from the path name
using this method.

<path-to>/foo.h ->

 #ifndef FOO_H
 #define FOO_H

 #endif

Fix docs formatting of all checks using with the HeaderFileExtensions
option.

llvm-header-guard is an alias for bugprone-header-guard